refactor: moved TokenIssueService.java to TokenIssueUtil.java
